+++ src/lib/fragments/AgGrid.react.js
@@ -66,6 +66,9 @@ const RESIZE_DEBOUNCE_MS = 200;
 // Rate-limit for updating columnState when interacting with the grid
 const COL_RESIZE_DEBOUNCE_MS = 500;
 
+// Time between syncing cell value changes with Dash
+const CELL_VALUE_CHANGED_DEBOUNCE_MS = 100;
+
 const xssMessage = (context) => {
     console.error(
         context,
@@ -119,6 +122,7 @@ export default class DashAgGrid extends Component {
         this.onCellClicked = this.onCellClicked.bind(this);
         this.onCellDoubleClicked = this.onCellDoubleClicked.bind(this);
         this.onCellValueChanged = this.onCellValueChanged.bind(this);
+        this.afterCellValueChanged = this.afterCellValueChanged.bind(this);
         this.onRowDataUpdated = this.onRowDataUpdated.bind(this);
         this.onFilterChanged = this.onFilterChanged.bind(this);
         this.onSortChanged = this.onSortChanged.bind(this);
@@ -924,20 +928,44 @@ export default class DashAgGrid extends Component {
         node,
     }) {
         const timestamp = Date.now();
+        // Collect new change.
+        const newChange = {
+            rowIndex,
+            rowId: node.id,
+            data,
+            oldValue,
+            value,
+            colId,
+            timestamp,
+        };
+        // Append it to current change session.
+        let pendingChanges = this.state.cellValueChanged;
+        if (typeof pendingChanges === 'undefined' || pendingChanges === null) {
+            pendingChanges = [newChange];
+        } else {
+            pendingChanges.push(newChange);
+        }
+        this.setState({cellValueChanged: pendingChanges});
+    }
+
+    afterCellValueChanged() {
+        const {cellValueChanged} = this.state;
+        // Guard against multiple invocations of the same change session.
+        if (
+            typeof cellValueChanged === 'undefined' ||
+            cellValueChanged === null
+        ) {
+            return;
+        }
+        // Send update(s) for current change session to Dash.
         const virtualRowData = this.virtualRowData();
         this.props.setProps({
-            cellValueChanged: {
-                rowIndex,
-                rowId: node.id,
-                data,
-                oldValue,
-                value,
-                colId,
-                timestamp,
-            },
+            cellValueChanged: cellValueChanged,
             virtualRowData,
         });
         this.syncRowData();
+        // Mark current change session as ended.
+        this.setState({cellValueChanged: null});
     }
 
     onDisplayedColumnsChanged() {
@@ -1337,7 +1365,11 @@ export default class DashAgGrid extends Component {
                     onSelectionChanged={this.onSelectionChanged}
                     onCellClicked={this.onCellClicked}
                     onCellDoubleClicked={this.onCellDoubleClicked}
-                    onCellValueChanged={this.onCellValueChanged}
+                    onCellValueChanged={debounce(
+                        this.afterCellValueChanged,
+                        CELL_VALUE_CHANGED_DEBOUNCE_MS,
+                        this.onCellValueChanged
+                    )}
                     onFilterChanged={this.onFilterChanged}
                     onSortChanged={this.onSortChanged}
                     onRowDragEnd={this.onSortChanged}